Most of the pieces complete!
I've just started moving the pieces in to place, and I'm
committing before I get too far along in case I make a mess of
things! But I've got all of the major functions written; I'm only
lacking some of the pretty and fancy stuff. (Not that this
program will ever be pretty or fancy!)