Stepped back a bit
I undid my changes to the R_n_D files related to organizing them
for the final product. I realized this was silly: I have all of
these excellently working functions and programs and it would be
much better (and easier!) to keep them as is and simply copy them
piece by piece to the appropriate places.