It's done! Fully functional anyway...
The game is fully functional. The messages and prompts still
have just placeholder text, but it's a simple change of a
macro to set these to whatever I want. Same goes for the window
size and placement; no need to mess with the board, but the
user name prompt and the message pop-ups will probably want to
be bigger and located in a different place; oh and definitely
play with the colors but again it's set up so that all remaining
modifications are simple macro changes. The only exception is
the handling of the end of a game, but I defined functions to
handle these incidents (a won game or a stalemate) and these can
be simply redefined to have a different response to these events.